+++ changelog.md
@@ -10,7 +10,8 @@
 - Add player profile pictures to high scores dialog
 - Rename ``LoadingIndicator`` to ``LoadingOverlay`` and increase its z-index to match dialog
 - Show player profile picture in members area instead of generic icon
-- Fix bug: TURTLE10 - Reset personal best on logout
+- Fix bug [TURTLE-10]: Reset personal best on logout
+- Fix bug [TURTLE-11]: Do not show login dialog if authenticated
 
 # 0.7.4 Beta - 14/08/2025 14:10 UTC+1